Add interactive manpage exploration script.
If no term is provided, then just list all of the files in the
subdirectories that have a file matching "man.*\.\d" (like man.3). If a
term is provided, then use Ag to search for the term only in files with
name path given above. Hitting enter will display the man page, and q
will return back to the interactive man page selector.